hostapd: Reconfigure sm and primary auth for assoc link
With ML reconfigure link removal support, any link can be removed from
the MLD. If the link that is getting removed is happened to be the assoc
link id for the connected station, then it has to be updated to have sm
and primary wpa auth to the next available partner link.

Also Currently when there is only one BSS, the ML reconfigure link
removal support disables the removed interface. This causes a failure
when attempting to add the same interface back to the AP MLD
because the interface object is not freed while removing it.

Add changes to reconfigure the state machine and the wpa_auth for the
associated station from next available partner hapd and ensure the
removed interface object is freed and the other interface objects
are properly handled.

Change-Id: Ia59ede05ef7e358342421f4f9919cc44ea6fb6b1
Signed-off-by: Manish Dharanenthiran <quic_mdharane@quicinc.com>
Co-developed-by: Yuvarani V <quic_yuvarani@quicinc.com>
Signed-off-by: Yuvarani V <quic_yuvarani@quicinc.com>