hostap: Fix interop issue
Currently, when 5 GHz is configured on 240 MHz bandwidth,
Support For 320 MHz in 6 GHz flag in EHT PHY Capabilities Information
gets enabled. This causes an interop issue in some 3rd party station.
Reset the flag if the channel is not a 6 GHz channel and the channel
width is not 320 MHz.

Fixes: Id5f982f725 (Fix MCS NSS map in beacon for BW less
    than of equal to 80MHz)

Change-Id: Ie7fd7de11f03b7f8704856cccf571c486dd0ccef
Signed-off-by: Hari Naraayana Desikan Kannan <quic_hnaraaya@quicinc.com>
(cherry picked from commit b03a4b3dafcf0f95e4ce3576eedecd2351c484e6)