hostapd: Fix PSD calculated for TPE1 in SP power mode
The TPE1 advertised in the beacon of an AP operating in the 6 GHz band
contains PSD power values. When the AP is operating in SP power mode,
the PSD power value is determined as the minimum of the PSD value
specified in the regulatory rules and the PSD value derived from the AFC
response, minus 6dBm. However, the PSD value should not be directly taken
from the AFC response. Instead, it should be calculated using the EIRP
power provided in the AFC response.

To implement this, modify the get_psd_for_chan_idx API to fetch EIRP
power values for each 20 MHz sub-channel and calculate the PSD power
value accordingly from this.

Tag: 6G_REGULATORY
Fixes: I13e7180a5b9 ("hostapd: Add AFC power compliant TPE for
		     Standard power AP")

Change-Id: Id1aa3b7678d0c79134ff82d5d3a28366d7b6e65b
Signed-off-by: Amith A <quic_amitajit@quicinc.com>