hostapd: Fix puncturing bitmap for 5G 240
Currently, when AP is configured in 5GHz 240MHz and 40MHz
puncturing is applied, the bitmap is not present in EHT
operation IE.

This is because for all 240 puncturing cases, the bitmap is
present in a separate vendor IE. But the puncturing bitmap
should be present in EHT operation if it has puncturing within
first 160MHz.

Fix this issue by updating the puncturing bitmap of first 160MHz,
if present, in the operation IE for 5GHz 240MHz.

Fixes: I5eb3fdeda77d (hostapd: Fix association issue with EHT320 and puncturing bitmap)

Patch-depdendency: r22-hostapd-fix-EHT320-with-puncturing-association-issue.patch

Patch-work:

Feature: PUNCTURING

Change-Id: Iae38261c5097a0eb8f9437935e371ab48e8eb214
Signed-off-by: Mohan Kumar G <quic_mkumarg@quicinc.com>