UPSTREAM: HOSTAPD: Fix hostapd crash when sta reassociates in legacy mode.
Currently, when a STA tries to associate as an ML client and
the association fails, re-associating the same STA in non-ML mode
causes the link to be removed. When __check_assoc_ies fails, the link
is freed again, leading to a hostapd crash as wpa_auth is already freed.

To address this issue, treat the STA as an ML STA when processing the
associated link.

Sometimes when STA re-tries new association with AP, the AP might use
the old sm to process the STA association. During this time, the sm
may not be cleared from its older values. it leads to having
incorrect data values in the sm structure causing connection failure.

Reset the number of affiliated link count in STA state machine
when assoc request received from STA. Also clear out the wpa_auth
reference in the sm. The sm structure gets updated with new
values upon current association.

Additionally, update linkId only if the associated link ID matches
when an auth request is received

Fixes: Ia59fc09ae0c6 ("hostapd: Reset number of affiliated link count
                     in Assoc request")

Fixes: I4bf2695b4daf ("hostapd: handle ML Secured associations properly")

Change-Id: I3e2c1eae8d293ac3accfb0a839edd84a2807a565
Signed-off-by: Arunpandi Kannan <quic_arunpand@quicinc.com>
Signed-off-by: Hari Naraayana Desikan Kannan <quic_hnaraaya@quicinc.com>
(cherry picked from commit ec21529d1cf92ee31430ad6bd63314f4d78224b3)
Signed-off-by: Manish Dharanenthiran <quic_mdharane@quicinc.com>