UPSTREAM:PROTOCOL:HOSTAPD: roaming fails with FT-Over-Air scenario
Station fails to roam with FT-Over-Air scenario, this is because of
failure mic validation at target AP and station and link address got
overwritten when STA exist in reassoc_req.

Pass the proper input for mic computation to make the verification
success at target AP and station. Also, the overwrite link address with
mgmt->sa only when the STA is not authorized and MFP is not done for STA.

Fixes: I3ce2a50e546 ("hostapd: QSDK patches update for package upgrade 13.0")

Change-Id: I27bdfffed8cf82bb09e7e2edbe33dee91d812839
Signed-off-by: Karthikeyan Kathirvel <quic_kathirve@quicinc.com>