UPSTREAM:HOSTAPD: Fix aid indexing in hapd->sta_aid array
The hapd->sta_aid array holds the AIDs (Association IDs) of all associated
STAs. When a STA disconnects, the AID associated with it is reset. When a
STA associates with an AP, its AID is set in the hostapd_get_aid function,
which is called from handle_assoc.

For instance, if this is the first STA associating with the AP, the AID in
sta_info is set to 1, as AIDs should start from 1. Consequently, the 1st
bit of the hapd->sta_aid[] corresponding to the STA is set. When the STA
disconnects, this set bit is reset in the ap_free_sta API. However, the bit
to be reset is calculated as BIT(sta->aid - 1), assuming that the indexing
starts from 0. This incorrect calculation results in resetting the 0th bit,
while the bit that was initially set was the 1st bit. This leads to
failures in subsequent associations, as the AID would be calculated as 0 in
the next association attempt.

To fix this issue, avoid subtracting 1 from sta->aid so that the correct
bit is calculated for setting and resetting the hapd->sta_aid[] value.

Fixes: I3b1e1d6f200("hostapd: Add multi MBSSID group support")

Change-Id: I3c5c99701fc6dfe7c8d08d99574566de8b6991ea
Signed-off-by: Amith A <quic_amitajit@quicinc.com>