UPSTREAM:HOSTAPD: Fix center frequency setting after CAC completion
When CAC is completed, the centre frequency is set based on the
channel and width where CAC is completed, assuming the AP will move
to that channel. However, for background CAC, no channel switch occurs,
leading to incorrect configurations if the centre frequency is changed.

Set the starting frequency in the configuration only if the CAC completion
is successful and channel switch is triggered. Do not update the center
frequency for background CAC. Update the center frequency only when CAC
is completed and the interface is down (CAC waiting period).

Change-Id: If88b46e79d7ee13dbc2c3d85c42207879b507aee
Signed-off-by: Hari Naraayana Desikan Kannan <quic_hnaraaya@quicinc.com>