QSDK: hostapd: Fix KW issue for attr might be NULL
Currently, In nl80211_vendor_cmd_if_offload_type() function, attr is
assigned a return value of a function which can be NULL but the code
does not check for NULL pointer value and proceeds to dereference it.
Fix this by checking attr for NULL value.

Fixes: bd6b598295e6 (QSDK: hostapd: Add support for configuring
interface offload type per netdevice)

Patch-depdendency: None

Change-Id: I692162d159630e4aba840909a2503c14f3f4e062
Signed-off-by: Aditya Rajendra Gawali <quic_agawali@quicinc.com>