UPSTREAM:HOSTAPD: Fix the PSD values filled in TPE IE 1
Currently, in the PSD TPE IE, the PSD value for each sub-channel is
calculated by fetching the corresponding 20 MHz EIRP and then converting
it to the PSD value using the formula:
	PSD = EIRP - 10log10(20)

In scenarios where the 20 MHz EIRP values of the sub-channels differ,
this calculation yields different PSD values for the sub-channels.

To ensure consistency, instead of using the 20 MHz EIRP value for each
sub-channel, use the EIRP value for the current operating bandwidth.
Therefore, modify the formula to calculate the PSD value as follows:
	 PSD = EIRP - 10log10(Current_BW)

This approach ensures that the PSD values are consistent across all
sub-channels by using the EIRP value for the entire operating bandwidth
rather than individual 20 MHz segments.

Tag: 6G_REGULATORY
Fixes: Id1aa3b7678d ("UPSTREAM:HOSTAPD: Fix PSD calculated
		     for TPE1 in SP power mode")

Change-Id: I81d0347ee96358057cd344a320ec1b56f8649b5a
Signed-off-by: Amith A <quic_amitajit@quicinc.com>