hostapd: Remove Wide Bandwidth Element for 20 MHz
Currently, the wide bandwidth element is present in the channel switch
wrapper when the maximum non-disabled bandwidth is more than 40 MHz.

According to IEEE P802.11-REVme/D4.0 section 11.38.4, the wide bandwidth
element must be present when switching to a 40 MHz channel width.
Conversely, it must not be present when switching to a 20 MHz channel width.

Fix the issue by removing the wide bandwidth element in channel
switch wrapper element for 20MHz.

Change-Id: Ic3e8a66633cc752650f1e967b8203b8611a31602
Signed-off-by: Hari Naraayana Desikan Kannan <quic_hnaraaya@quicinc.com>
Signed-off-by: Mohan Kumar G <quic_mkumarg@quicinc.com>