QSDK:HOSTAPD: Fix KW issue for attr might be NULL
Currently, In nl80211_vendor_cmd_if_offload_type() function, attr is
assigned a return value of a function which can be NULL but the code
does not check for NULL pointer value and proceeds to dereference it.
Fix this by checking attr for NULL value.

Fixes: 3fec286f1cf3 (QSDK: hostapd: Add support for configuring
interface offload type per netdevice)

Patch-dependency: None

Change-Id: Ia64881fd561fd1ebef0c94e176ed06353b1eeb6c
Signed-off-by: Aditya Rajendra Gawali <quic_agawali@quicinc.com>