UPSTREAM:HOSTAPD: Call hostapd_drv_set_6ghz_pwr_mode for all the BSS
When are there are multiple 6 GHz AP interfaces in SLO, and when
set_6ghz_pwr_mode command is executed for the first interface,
the iw command's output shows the incorrect power mode for the other
APs' interfaces.

This is because, the command with the first interface does not apply
for the other APs' interfaces.

To fix this issue, call hostapd_drv_set_6ghz_pwr_mode for all the
BSS in the iface, instead of the BSS coming in the CLI command.

Tag: 6G_REGULATORY
Fixes: I0a0324ac226 ("hostapd: Add power mode to channel switch command")
Fixes: I29fee09190f ("hostapd: Add hostapd cli support to switch 6 GHz power mode")

Change-Id: I3b8f2ec9e19c70b3d53458bf06d60f07ef5109d4
Signed-off-by: Hariharan Basuthkar <quic_hbasuthk@quicinc.com>