UPSTREAM:PROTOCOL:HOSTAPD: Fix deletion of RNR info in partner links
With MBSSID grouping enabled by default, when Tx vap is removed all
the Non-Tx vaps in the group (say wlan0 (Tx vap),wlan1 (Non-Tx),
wlan2 (Non-Tx),wlan3 (Non-Tx) are in a group) will also get removed.
The respective partner links (2 GHz/ 5 GHz) should be informed
regarding the removal of the Tx and Non-Tx vaps which in turn
should advertise the existing vaps in their respective RNR IE.

However this is not happening, instead when Tx vap (say wlan0)is
removed in 6G, its partner (2 GHz/5 GHz -wlan0) removes only the
respective Tx vap and the Non-Tx vaps (6G - wlan1, wlan2, wlan3)
info is still present in the RNR IE which is not expected. The
same is happening with the other Non-Tx vaps which removed along
with Tx vaps. The partners (2 GHz/5 GHz) of other groups (say
wlan4,5,6,7 form group2) still advertise the removed Tx and Non-Tx
vaps (group 1) in their RNR which is also not acceptable.

Fix this by advertising the removal to all the partners in the
interface via set beacon instead of advertising to their
respective partner alone.

Fixes: Ie2ba6a8220a7 (hostapd: Add Multi-Link Reconfigure link
removal support)

Patch-work: None
Patch-dependency: None

Change-Id: I0da6acc2db6708e1bb1233fb8149023e7e0ed8bd
Signed-off-by: Dhanavandhana Kannan <quic_dhanavan@quicinc.com>