UPSTREAM:HOSTAPD: Set complete list of Nontxbssid profile bit to 1 with MBSSID grouping in 6GHz.
In 6G Complete list of Nontxbssid profile bit in Extended
capability IE is 0 but should be 1 since profile periodicity
is 1 with MBSSID grouping.

To fix the issue, in case of MULTI_MBSSID_GROUP_ENABLED complete
variable is set to true in ieee802_11_build_ap_params function.

Patch-depdendency: None

Patch-work: None

Fixes : 74a5a6d5cbc (hostapd: Add multi MBSSID group support)

Change-Id: I9f1f4244f2f797a08884dafe12d3415ffb083d79
Signed-off-by: Pavithra Ganesan <quic_pavigane@quicinc.com>