UPSTREAM:PROTOCOL:HOSTAPD: Fix advertise RNR info of the specified group alone in FILS discovery frame
In FILS discovery request frame in 6 GHz standalone mode, RNR IE
contains info of all the vaps from all the groups. But the expectation
is when grouping is enabled, RNR should advertise the info of the vaps
of that particular group alone as each group has separate Tx vap.

Fix by enabling RNR to advertise the info of vaps of the respective
group alone when grouping is enabled in 6 GHz standalone mode.

Fixes: I3b1e1d6f2006(hostapd: Add multi MBSSID group support)

Patch-dependency: r45-001-hostapd-Add-Multi-MBSSID-group-support.patch
Patch-work: None

Change-Id: I75f60807b486a52ebf73a5320a59961be1207eea
Signed-off-by: Dhanavandhana Kannan <quic_dhanavan@quicinc.com>