UPSTREAM:PROTOCOL:wpa_supplicant: fix NULL pointer crash during ML association
When association frame is received from an MLD AP, the station
checks if all links are associated via the is_all_links_associated()
function.

Inside this function freq_list is accessed without verifying its
availability, as freq_list is not mandatory in the station configuration.
This can lead to a NULL pointer crash in wpa_supplicant.

To address this issue, allow the all-link association check
only when freq_list is present in the station. Otherwise,
the station should associate with the first received
association frame.

Fixes: 1cd6199de285 ("wpa_supplicant: Add changes to reassociate if associated links is lesser than requested link)
Patch-dependency: None
patch-work: None

Change-Id: I17c956f5711cf26964c66ba8d50b7045a88827e3
Signed-off-by: Aaradhana Sahu <quic_aarasahu@quicinc.com>