QSDK:PROTOCOL:wpa_supplicant: fix null pointer crash in wpas_scan_for_rnr_entries
When multi-link station interface is removed using interface remove command
in wpa_supplicant, a null pointer crash is seen in wpas_scan_for_rnr_entries().

This is because, wpas_scan_for_rnr_entries eloop timer is scheduled during
association as it was connected in single link but when we disable that
interface, since eloop work is not cancelled, eloop tries to execute
wpas_scan_for_rnr_entries and crashes with null data in eloop_ctx.

Add eloop_cancel_timeout for scan_rnr in wpa_supplicant_cleanup().

Change-Id: I18113697bd81674c8bf52d3f175407b4f9daccbe
Signed-off-by: Harshitha Prem <quic_hprem@quicinc.com>