UPSTREAM:PROTOCOL: hostapd: add support to use bitmap to get link ids
Currently, link IDs are assigned based on a counter
stored in the hostapd MLD structure. With ML Reconfig
Link Removal, any link can be removed and, when re-added,
is assigned a new link ID. This process can eventually exhaust
all available link IDs.

To address this issue, use a bitmap instead of a counter.
A value of 1 at a given index indicates that the link ID
is available, while a value of 0 means it is currently
in use. This method simplifies tracking link IDs,
allowing them to be reused when necessary.
Also, retain the link ID of the disabled interface,
to ensure that the disabled interface link ID is not
utilized during dynamic link addition.

Patch-dependency:
r32-005-hostapd-Handle-link-reconfig-event-started-and-compl.patch
r44-001-hostapd-ML-reconfiguration-link-addition-support.patch

Patch-work: None

Change-Id: I9531e39bf391139baa2f9c1d6f37539065a23dbd
Signed-off-by: Yuvarani V <quic_yuvarani@quicinc.com>