UPSTREAM:PROTOCOL: Hostapd: Fix auto channel hostapd crash
When AP is configured with auto channel in 2 GHz and 5 GHz bandwidth
is more 20 MHz, Hostapd crashes due to the NULL pointer access in
hostapd_event_get_survey().

The ACS is activated for the 2.4 GHz band, initiating a scan with a 20-second
timeout. The 2.4 GHz band updates the scan context accordingly.
When the 5 GHz band is initialized, an HT40 scan is attempted due to the
bandwidth exceeding 20 MHz. However, the driver returns a "Resource busy"
message because of the ongoing ACS scan. An HT40 scan starts an eloop
timer for a retry. The HT40 scan retry timer is triggered, and in the meantime,
ACS scan results for the 2 GHz band are received. The HT40 scan is then
prioritized, and the driver initiates a scan, updating the scan context
for the 5 GHz band.

When the ACS event survey is processed, it uses the updated scan context.
During the survey event processing, the scan callback for the partner
link (2.4 GHz ACS) is handled, which leads to a crash.

Fix by intiating the HT scan after the ACS triggered scan is completed.

Patch-depdendency: None

Patch-work: None

Feature: SCAN

Change-Id: I5fa71a118f3455e73d370886dc6c22a8d2913782
Signed-off-by: Karthik <quic_karm@quicinc.com>