UPSTREAM:HOSTAPD: Fix radar event processing
When a radar event is received in an AP MLD operating on a DFS channel,
nl80211_radar_event() iterates over all the BSSs available in drv to
find a link matching the frequency of the event. If a link match is
found, the radar handler function tries to switch to a new channel with
the same bandwidth. In case no valid channels are available it disables
and re-enables the interface, reallocating the drv BSSs. However, the
loop in nl80211_radar_event() function continues to access the old
deallocated BSSs' address in the next iteration, causing a crash.

Since the radar handler function handles the event for all BSSs in an
interface, there is no need to call it again once a link match is found.
Hence, fix this issue by exiting the loop after calling the handler if a
link match is found for the radar event.

Also, since the loop already checks all the BSSs, remove the handler
present before the loop.

Patch-depdendency: None

Patch-work: https://w1.fi/cgit/hostap/commit/?id=338e79b6f8d452a638da5d746a3fae0f5b5b5bd0

Feature: RADAR

Change-Id: I10287ecc295ced70844f6a5e10c91128424b6da1
Signed-off-by: Mohan Kumar G <quic_mkumarg@quicinc.com>