UPSTREAM:HOSTAPD : Rectify issue in 6 GHz 20 Mhz channel switch
20 MHz Channel switch failure is observed in the 6 GHz link of an
MLO AP.

The kernel reports an error in the nl80211_policy with the CSA IEs
sent in the nl80211_cmd_switch_channel. The bandwidth indication
element is not added to the beacon template in case of 20 MHz channel
or in absence of puncturing.

In hostapd, within hostapd_eid_chsw_wrapper, during a 20 MHz channel
switch, both wide bandwidth and bandwidth indication elements are not
present. But the tag number and length of channel switch wrapper
element are added. The eid is returned without updating the
eid_len_offset. This causes the corruption of beacon_ie_len that is
then sent to the kernel in nl80211_cmd_switch_channel.

To fix this issue, do not add channel switch wrappper element if both
wide bw and bandwidth indication subelements are not present.

Patch-dependency: r41-hostapd-Update-correct-width-in-wide-bandwidth-chann.patch
Feature: CHANNEL_SWITCH
Fixes: I339d94aef07 ('UPSTREAM:PROTOCOL:HOSTAPD: Update correct width in wide bandwidth channel switch
')

Change-Id: I55b2d2ce67b83986a4b068662cfa20fb25cb4066
Signed-off-by: Hariharan Basuthkar <quic_hbasuthk@quicinc.com>