QSDK:CP: Handle EHT240 MHz for 5 GHz band
Currently, when AP+MESH is configured in EHT240 MHz, AP interfaces
are not coming up via ucode due to incorrect center frequency calculation.

For EHT240, the centre frequency index is always 130 and can be brought
up only in channel 100-144. Hence, add changes to set center frequency index
as 130 for EHT240 in freq_info handler of ucode.

Change-Id: Ia7702c4a456a85a335fc553d2a39a78547795f20
Signed-off-by: Harshitha Prem <quic_hprem@quicinc.com>