QSDK:PROTOCOL: hostapd: STA disassociates when steering to other band
When the station tries to steer to other band in a same MLD, STA gets
pruned by AP this is because there was a change got reverted in OpenWrt
for other vendor driver specific fix.

Revert the other vendor specific change in OpenWrt to fix this issue for ath
drivers.

Fixes: 5aab43f93388 ("hostapd: slightly clean up patches")

Change-Id: I9ecd63c91a06cf904e19aabbb0894848eafca1df
Signed-off-by: Karthikeyan Kathirvel <quic_kathirve@quicinc.com>