UPSTREAM:PROTOCOL: Free STA Associated to Partner BSS on Link Disable
Currently, when a link is disabled and a STA is associated
as a partner link to the current link, a broadcast deauth is sent,
and the STA gets disconnected. However, since the STA is connected
only as a partner link (non-assoc link), the STA reference is not
deleted and is only removed after the maximum inactivity time.

This cause an issue where the STA tries to associate with same or
another link after deauth, but fails because the STA reference remains.

To address this issue, when a link is disabled, check if the STA
is a partner link of the currently disabled BSS and remove the STA
reference, allowing the STA to retry association.

Patch-depdendency: None

Patch-work: Free STA Associated to Partner BSS on Link

Feature: MLSTA

Change-Id: Idaaa9ae71f7447358517f5158860bf97301891b8
Signed-off-by: Hari Naraayana Desikan Kannan <quic_hnaraaya@quicinc.com>