UPSTREAM:HOSTAPD:Revert "QSDK:HOSTAPD:Revert hostapd: fix dangling pointer access in current hwmode"
Channel + power switch failure is observed on the 6 GHz high radio of a
RDP (RDP 464) which supports both 6 GHz low and high radio. This is
observed when power mode is configured with 0 or 1 after AFC is complete.

The iface->current_hw_info is not updated when
NL80211_CMD_WIPHY_REG_CHANGE event is received from the kernel.

To fix this issue, bring back the patch
"hostapd: fix dangling pointer access in current hwmode" .

This reverts commit d623b4d6e3dd709e265f06810361ca79b47586bc.

This change is available in upstream hostapd - https://w1.fi/cgit/hostap/commit/?id=acc35945a987dddf921903a7d0c5d4279970e7bd and it needs to be removed in the next Hostapd upgrade.

Change-Id: Ib7ec3f5713119ec4ae9e8758d33e2f859dc7df5a