QSDK:HOSTAPD: Avoid CAC during BW expansion on 5 GHz 240 MHz
In an ETSI country like Qatar, when a WDS AP is associated with a WDS
Station on the 5 GHz band at 240 MHz on channel 100, the bandwidth is
reduced to 160 MHz and then expanded back to 240 MHz.
Disconnection is observed during the bandwidth expansion to 240 MHz.

This occurs because hostapd_is_dfs_overlap returns a positive value
for res, as the half-width (160 MHz) overlaps with channel 149.
Since channel 149 is a DFS channel in Qatar, this causes the AP to
undergo CAC, leading to the station’s deauthentication from the AP.

To fix this issue, in hostapd_is_dfs_overlap, check if every
subchannel of the current AP’s bandwidth has undergone CAC.
If yes, proceed with the channel switch for bandwidth expansion;
otherwise, undergo CAC.

Feature: DFS
Fixes: I25eea7829de ("q02-046-hostapd-Add-5GHz-240MHz-support.patch")
       I55ef67e8006 ("q02-046-hostapd-add-support-to-disable-channel-switch-during.patch")

Change-Id: Iba686f6bdc70a4cd214889670a1d2a8fc111d46c
Signed-off-by: Hariharan Basuthkar <quic_hbasuthk@quicinc.com>