QSDK:HOSTAPD: Get chwidth from vht_oper_width for VHT CAP
When a WDS station is connected with a RootAP on channel 100 in 160
MHz bandwidth, and the AP switches to channel 100 in 240 MHz,
the station moves to 80 MHz bandwidth and fails to switch to 240 MHz.

On the station side, the channel switch to 240 MHz is possible only
if the station first moves to 160 MHz and the vendor IE is looked up
for the puncturing pattern, for moving to 240 MHz. However, in this
case, the station moves to 80 MHz bandwidth due to the absence of
the 160 MHz support bit in the VHT CAP IE of the RootAP’s beacon
just after the beacon frame with CSA.

The VHT capabilities present in the RootAP’s beacon are filled in
hostapd_eid_vht_capabilities. In hostapd_eid_vht_capabilities,
the channel width is filled using the API hostapd_get_oper_chwidth.

During the channel switch, in hostapd_event_ch_switch, the channel
width is filled as 320 in conf->eht_oper_width and 160 in
conf->vht_oper_width. Since hostapd_get_oper_chwidth fetches
conf->eht_oper_width with 320, in hostapd_eid_vht_capabilities the
160 support capability bit is not filled.

To fix this issue, in hostapd_eid_vht_capabilities, fill the channel
width with conf->vht_oper_width.

Feature: CHANNEL_SWITCH
Fixes: If2629d562015("Fix the station bandwidth inconsistency issue")

Change-Id: I16bdd5fae7fe06bae885898684fcf11cba007c6a
Signed-off-by: Hariharan Basuthkar <quic_hbasuthk@quicinc.com>