UPSTREAM:PROTOCOL:hostapd: Fix re-addition of Tx BSS in a MBSSID Set
When a BSS corresponding to a Transmitted BSSID is removed in a MBSSID
Set, all the BSS in the group(including Tx BSS and Non-Tx BSS) will
get removed. When the removed BSS is added again, it should get added
as Tx BSS in the MBSSID set but it is not happening as the BSS added
is not set as Tx BSS and it results in the failure of ML link addition
as it results in crash when Tx BSS is accessed.

To fix this ensure that when a BSS is added in a MBSSID set and Tx BSS
is NULL, mark the added BSS as Transmititng BSS of the MBSSID Set.

Fixes: I3b1e1d6f2006 (hostapd: Add multi MBSSID group support)
Patch-dependency: r45-001-hostapd-Add-Multi-MBSSID-group-support.patch
Patch work: None

Change-Id: If0b84091ac042eff0ada2b30a03e3ab56a592056
Signed-off-by: Dhanavandhana Kannan <quic_dhanavan1@quicinc.com>