QSDK:HOSTAPD: Fix puncture bitmap in vendor IE of CSA beacons
When an AP is configured with disable_csa_dfs and radar is injected,
no CSA is triggered from the AP, and it waits for a channel switch
command from the user. When the user triggers the channel switch
command to reduce the bandwidth, the CSA beacons are found to lack
the puncturing bitmap in the vendor IE. This occurs because the
puncturing bitmap is being cleared during radar handling. This
behavior should not occur during CSA with puncturing
(i.e., reducing bandwidth), as the vendor IE in the CSA beacon would
not have the puncturing bitmap, leading to STA disconnections.

To resolve this issue, avoid resetting the puncturing bitmap before
channel switch is complete.

This ensures that the current puncture bitmap is retained in the Vendor IE of the CSA beacons. Reset the puncture bitmap when there is a disable-enable interface sequence for switching to any other DFS channel.

Feature: RADAR

Change-Id: I4906b1a3b40cfe247fd1478b47786b234657e1a9
Signed-off-by: Amith A <quic_amitajit@quicinc.com>