UPSTREAM:CP:hostap: Fix scan frequency list in Mesh mode
In Mesh mode, before the vdev gets created it tries to scan with a
list of frequencies.

Previously when the frequency list to scan is not filled in the scan
request, the mac80211 triggers scan for all the band and in this flow
the scan vdev was not getting deleted and so key installation was
happening in wrong band due to which ping was not successfully in Mesh
mode. This issue is fixed by sending the whole band frequency to scan
when frequency list is not provided in the supplicant config file.
However, with respect to latest scan design, the scan vdev created
during scan phase will get deleted once the scan status arrives from
target. So these changes are no longer required.

This scan frequency list now comprises of all the frequencies
in that band even-though the frequency list(freq_list) is mentioned
in the wireless config file.

Consider, a device has 6 GHz band split into two as 6 GHz Low
(5925 MHz - 6425 MHz) and 6 GHz High (6585 MHz - 7125 MHz) band in
two SoCs, then this whole frequency list from wpa_supplicant triggers
scan in both the SoCs which is incorrect.

This behaviour causes the issue in the following case. When 6 GHz Low
is configured with max vdevs and 6G GHz High is configured with max-1
vdev and the remaining one interface is a mesh, then during scan for
that mesh vdev, scan is getting triggered for both the SoCs.
This results in an error because there are no free vdevs left in 6 GHz
Low Soc as scanning needs a vdev.

Fix this issue by ensuring the whole band frequency list is not sent,
regardless of the frequency list's presence in the supplicant config
file.

Patch-depdendency: 955-wifi-ath12k-Split-scan-request-for-split-band-device.patch

Change-Id: Ib1de70fb551ab542bc5d606d207905e373dbe5dc
Signed-off-by: Nagarajan Maran <quic_nmaran@quicinc.com>