UPSTREAM:PROTOCOL:hostapd: Enable removal of interface only in SLO or legacy case in WDS mode
Currently when primary TQM link is removed in WDS connection
with primary UMAC migration enabled, the whole VLAN interface
is brought down beacuse the sta flag has been enabled only
for primary link and when primary TQM is removed, it results in
bringing down the VLAN interface which is not expected.

Fix is to initiate the removal of interface only when partner
links do not exist or in legacy or SLO cases.

Patch-dependency: r03-029-hostapd-Enable-WDS-AP-MLO-Support.patch
Patch-work: None

Change-Id: I45ed4eb06715971752a103d403a37cc3860be1c3
Signed-off-by: Dhanavandhana Kannan <quic_dhanavan1@quicinc.com>