UPSTREAM: PROTOCOL: Rectify CAC start issue on 5G low and high
In a 16 AP MLD and one  monitor vap config, there is an intermitent
5G low or 5G high link bringup failure observed on a 5G low - 5G high
supported RDP due to CAC start failure.

When scan trigger is done by hostapd for link 0 (low radio), bss->scan_link
is of link0. Since scan of link 0 is completed in kernel/mac80211,
-EBUSY is not returned by hostapd_driver_scan. Scan trigger happends is
done by hostapd for link 2 (high radio). Now, bss->scan_link is of link2.

Meanwhile, Scan result is received in the hostapd from the kernel
for low radio frequencies. But since bss->scan_link is of link2,
mld_link ctx is stored of that of link2 instead of link 0.
This leads to interface bringup of the high radio frequencies,
and CAC start is done, leading to scan in progress error from the kernel.

To fix this issue, start the driver scan within hostapd_driver_scan,
only if the scan is completed in the partner links. If scan is still in
progress, return -EBUSY. The callers of hostapd_driver_scan, which
are ieee80211n_check_40mhz and ap_ht40_scan_retry in this case, will
call ap_ht40_scan_retry after a timeout.

Fixes: I5fa71a118f3('UPSTREAM:PROTOCOL: Hostapd: Fix auto channel hostapd crash')
Patch-dependency: None
Feature: SCAN

Change-Id: Ic7bf05d0d0301624cba102a206630c67caa5b619
Signed-off-by: Hariharan Basuthkar <quic_hbasuthk@quicinc.com>