QSDK:CP:hostapd:ucode Add frequency validation before interface start
Add frequency validation for new requested frequency during
interface start.

In WDS STA (5 GHz + 6 GHz) + repeater AP (2 GHz + 5 GHz + 6 GHz)
configured in MLO, some iterations of the STA connection result in
connection to only one link. During this scenario, the interface start
is called for all the disabled links of the AP with the associated
link frequency, causing two vdevs to be created in the same MLD
under the same radio. This commit adds frequency validation with the
channels information present in that radio, for the new requested
frequency during interface start.

Change-Id: If8ad80d92450ea212f061d45957a5e74421d63f8
Signed-off-by: Nagarajan Maran <quic_nmaran@quicinc.com>