UPSTREAM:PROTOCOL: hostapd: Updated started flag when the interface is stopped via some other userspace application
When any MLD/legacy interface is deleted through other userspace
network application like iw/ifconfig, kernel will intimate that through
STOP_AP event to hostapd. Currently, for this STOP_AP event, we will
set the interface to be stopped and incase of MLD, additionally we will
remove the link from the bss list. Incase of MBSSID case, if the deleted
interface is TX BSS, then the corressponding non-TX BSS will also get
deleted and recieve STOP_AP event similar to the orginally deleted
interface.

In current scenario, we will not updating the beacon of the partner
links/interface in this case, which leads to the deleted/stopped BSS RNR
information still present in the beacon as neighbor entry.

Avoid this by setting the disable flag to true for the stopped BSS,
refresh all the interface beacon and check whether the bss is disabled
or not before adding to the RNR/MBSSID element(s).

Change-Id: Ied4e44203c8e452cde01212a527c7fe92d0ddf4b
Signed-off-by: Manish Dharanenthiran <quic_mdharane@quicinc.com>