UPSTREAM:PROTOCOL: Fix legacy station assoc failure
Currently, when a station tries to associate as a multi-link (ML)
connection but fails to connect successfully
(station disassociated/deauthed and marked as not authorized),
we maintain the station context until the maximum inactivity timer
expires. If that particular station tries to re-associate as a legacy
station on a different link (non-associated link), the association
fails after EAPOL because we have not cleared the associated link
context. The existing code clears the link station context during
authentication only if the station is associated as a multi-link (ML)
connection.

To address this issue, clear the partner link for any station
association, provided the partner link is not authorized. This change
ensures that the station can successfully associate as a legacy
station on a different link without encountering association failures.

Patch-dependency: None

Patch-work: "Fix legacy station assoc failure"

Feature: ML STA

Change-Id: I88ea00c094d38a1906684c777e336b7f6c9311a4
Signed-off-by: Hari Naraayana Desikan Kannan <quic_hnaraaya@quicinc.com>