QSDK:HOSTAPD: Reset punct_bitmap in hostapd_dfs_start_channel_switch_cac
A 5 GHz SLO AP is configured on channel 100 in 240 MHz bandwidth goes
down, when radar is injected during CAC.

When a new random channel is selected for AP operation radar, and when
NL80211_CMD_START_RADAR_DETECTION is sent to kernel with the new random
channel, the kernel throws an error "Invalid channel definition". This
is because, the puncture pattern parameter is set to 0xF000 (which
corresponds to 320 MHz) while the new random channel has a bandwidth
of 160 MHz.

In hostapd_dfs_start_channel_switch_cac, the iface->conf->punct_bitmap
is not reset to 0. It has the puncturing pattern of the previous channel
(100 HT320 with 0xF000 puncturing pattern).

To fix this issue, reset iface->conf->punct_bitmap to zero in
hostapd_dfs_start_channel_switch_cac

Feature: RADAR, 5G240
Fixes: I4906b1a3b40 ("QSDK:HOSTAPD: Fix puncture bitmap in vendor IE of CSA beacons")

Change-Id: I1c7f760309b06ced27f4bf944929d105a61dd0ed
Signed-off-by: Hariharan Basuthkar <quic_hbasuthk@quicinc.com>