QSDK:PROTOCOL: hostapd: fix dropping frames when NL socket is congested
The burst of deauthentication frames and the deletion of STA entries are
causing congestion in the network layer. This congestion results in the
dropping of deauthentication frames at the kernel level, leading to STA
entries becoming stale.

Increase the RX socket limit in kernel, so that hostapd can allocate the RX
socket size as 512KB this makes kernel to send more frames, e.g receiving
burst of deauth frames to process STA deletion to make the NL congestion
less likely to occur.

Patch-work: None
Patch-dependency: None

Change-Id: Ifce21fcf37420e0a728ea270501629a75af40054
Signed-off-by: Karthikeyan Kathirvel <quic_kathirve@quicinc.com>