UPSTREAM:PROTOCOL:hostapd: Fix BSSID Index in Non-Transmitted Profile in MBSSID
Currently the BSSID Index field in the Multiple BSSID-Index element
in the non-transmitted BSSID profile corresponding to the
non-transmitted BSSID is set according the number of vaps being added
to the interface instead of number of vaps added in each MBSSID
set in the interface when MBSSID grouping is enabled.

Fix by setting the bssid index to a value between 1 and 2^(n)-1 when
the size of MBSSID set is 2^n as per IEEE Std 802.11-2020
section 9.4.2.73.

Patch-dependency: r45-001-hostapd-Add-Multi-MBSSID-group-support.patch
Patch-work: None

Fixes: I3b1e1d6f2006 (hostapd: Add multi MBSSID group support)

Change-Id: Ie03b91ef3484c5f626165e507d1e549845540177
Signed-off-by: Dhanavandhana Kannan <quic_dhanavan1@quicinc.com>