UPSTREAM:HOSTPAD : Rectify scan on 20 MHz and increase the scan retry timeout
When a 3 Link MLD AP configured with 6 GHz channel as auto, and 2 and 5
GHz channel configured with static channels, the 5 GHz link does not come
up.

When 6 GHz link is undergoing ACS, the 5 GHz link waits to undergo
HT40 scan. Scan retry happens for 14 attempts and after that 5 GHz link
tries to come up on 20 MHz and fails. In ap_ht40_scan_retry, the
secondary channel offset is set for 20 MHz, but the channel bandwidth,
and seg0 and seg1 frequencies are not updated to that of a 20 MHz
channel and as a result, the kernel returns an error.

To fix this issue, in ap_ht40_scan_retry, set the bandwidth 20 MHz, and
set the seg0 and seg1 frequencies to 0, to enable AP to come up on 20 MHz
bandwidth.

The eloop timeout for the 5 GHz scan retry is 1 second. Since ACS
scan on 6 GHz, makes the 5 GHz scan retry attempts to go until 14,
it makes the 5 GHz link to come up on 20 MHz. This behaviour is not ideal,
when the 5 GHz link is configured with bandwidths with 80 MHz or more.

To fix this issue, increase the eloop timeout from 1 to 5 in
ap_ht40_scan_retry.

Fixes: I5fa71a118f3('UPSTREAM:PROTOCOL: Hostapd: Fix auto channel hostapd crash')
Patch-dependency: None
Feature: SCAN

Change-Id: I4eb7d813306204af032e0bf461d631799b59961b
Signed-off-by: Hariharan Basuthkar <quic_hbasuthk@quicinc.com>