UPSTREAM:PROTOCOL:HOSTAPD: Fix set transmitted BSSID flag for every Transmitting BSSID in MBSSID Set
Before MBSSID grouping was introduced, the first BSS in the 6 GHz
acted as the Transmitting BSS, while the other BSSs remained as
non-transmitting BSS. Consequently, the transmitted BSSID flag in
the BSS Parameters subfield of the reduced neighbor report for 6 GHz
was set only for the first BSS in its partner beacons (say 2 GHz/5 GHz).

With the advent of MBSSID grouping, each MBSSID Set now includes a
Transmitting BSS. The transmitted BSSID field in the BSS parameters of
the Reduced Neighbor Report should be set for all Transmitting BSS's
in each MBSSID Set. However, the transmitted BSSID field in the Reduced
Neighbor Report is currently set only for the first BSS of the interface,
which is not the expected behavior.

The fix involves ensuring that the BSS advertised in the Reduced Neighbor
Report is a Transmitting BSS by verifying it against each MBSSID Set's
Transmitting BSS.

Patch-dependency: r45-001-hostapd-Add-Multi-MBSSID-group-support.patch
Patch-work: None

Fixes: I3b1e1d6f2006 (hostapd: Add multi MBSSID group support)

Change-Id: I0e0044c3dff2e69e85822b882011ade6cda64988
Signed-off-by: Dhanavandhana Kannan <quic_dhanavan1@quicinc.com>