UPSTREAM:PROTOCOL: hostapd: Fix setting power type to 255
In current implementation, while filling the CSA settings and generating
beacon, power_mode from csa_settings struct will be directly assigned to
he_6ghz_reg_pwr_type in hostapd_conf struct. power_mode is of type
integer as it will be set to -1 to begin with and have few checks while
getting the power mode from CLI or from kernel. Since,
he_6ghz_reg_pwr_type is of type u8, setting a negative value is
resulting in setting the 'reserved' bit in control field in HE operation
tag. Because of this, station(s) which are already associated is failing
to reserve the channel as the regulatory type sent from the AP is wrong.

Fix this by updating the he_6ghz_reg_pwr_type only when the power_mode
is greater than -1.

Patch-dependency: r31-hostapd-Add-power-mode-to-channel-switch-command.patch

Change-Id: Ia0f5bbd7f0ca08f5e2bf08e07576dcedb9670bd9
Signed-off-by: Manish Dharanenthiran <quic_mdharane@quicinc.com>